+++ app/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from sqladmin import Admin
 
 from config import i_config
-from includer.admins import admin_classes
+from admins import admin_classes
 from includer.routers import routers_all2
 from routers import routers_all
 from models.database import engine
